Add Docker

+++ src/main/java/com/jeeva/storemateapi/service/CustomerOrderService.java
@@ -1,9 +1,8 @@
 package com.jeeva.storemateapi.service;
 import com.jeeva.storemateapi.model.CustomerOrders;
-import java.util.Iterable;
 
 public interface CustomerOrderService {
     Iterable<CustomerOrders> getAllCustomerOrders();
     CustomerOrders saveCustomerOrder(CustomerOrders customerOrders);
-    CustomerOrders getCustomerOrderById(Long id);
+    CustomerOrders getCustomerOrderById(Integer id);
 }
